+++ src/main/scala/io/picnicml/doddlemodel/linear/typeclasses/LinearRegressor.scala
@@ -7,7 +7,7 @@ import io.picnicml.doddlemodel.typeclasses.Regressor
 trait LinearRegressor[A] extends LinearModel[A] with Regressor[A] {
 
   override protected def fitSafe(model: A, x: Features, y: Target): A = {
-    val wInitial = DenseVector.zeros[Double](x.cols + 1)
+    val wInitial = DenseVector.zeros[Float](x.cols + 1)
     copy(model, w = maximumLikelihood(model, xWithBiasTerm(x), y, wInitial))
   }
 }
