+++ src/test/scala/io/picnicml/doddlemodel/data/CsvLoaderTest.scala
@@ -11,14 +11,14 @@ class CsvLoaderTest extends FlatSpec with Matchers with TestingUtils {
   "Csv loader" should "load and encode data" in {
     val (x, y, featureIndex) = loadDummyCsvReadingDataset
     val xCorrect = DenseMatrix(
-      List(0.0, 0.0, 0.1, 1.1),
-      List(1.0, Double.NaN, 0.2, 1.2),
-      List(2.0, 1.0, 0.3, Double.NaN),
-      List(3.0, 2.0, 0.4, 1.4),
-      List(0.0, 0.0, 0.1, 1.1),
-      List(3.0, Double.NaN, 0.4, 1.4)
+      List(0.0f, 0.0f, 0.1f, 1.1f),
+      List(1.0f, Float.NaN, 0.2f, 1.2f),
+      List(2.0f, 1.0f, 0.3f, Float.NaN),
+      List(3.0f, 2.0f, 0.4f, 1.4f),
+      List(0.0f, 0.0f, 0.1f, 1.1f),
+      List(3.0f, Float.NaN, 0.4f, 1.4f)
     )
-    val yCorrect = DenseVector(0.0, 1.0, 2.0, 3.0, 0.0, 3.0)
+    val yCorrect = DenseVector(0.0f, 1.0f, 2.0f, 3.0f, 0.0f, 3.0f)
     breezeEqual(x, xCorrect) shouldBe true
     breezeEqual(y, yCorrect) shouldBe true
     featureIndex.names shouldBe IndexedSeq("f0", "f1", "f2", "f3")
